cloud/backend/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.18
click==8.1.7
cryptography==41.0.5
dnspython==2.4.2
ecdsa==0.18.0
pydantic-settings
email-validator==2.1.0.post1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fastapi==0.104.1
fastapi-mqtt==2.0.0
filetype==1.2.0
frozenlist==1.4.0
ftfy==6.1.1
gmqtt==0.6.12
h11==0.14.0
idna==3.4
joblib==1.3.2
lxml==4.9.3
Markdown==3.5.1
motor==3.2.0
msg-parser==1.2.0
multidict==6.0.4
nltk==3.8.1
numpy==1.26.2
olefile==0.46
openai==0.27.8
openpyxl==3.1.2
pandas==2.1.3
passlib==1.7.4
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==10.1.0
pyasn1==0.5.0
pycparser==2.21
pydantic>=2.3.0
pydantic_core==2.14.1
pymongo==4.6.0
pypandoc==1.12
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.23
pytz==2023.3.post1
regex==2023.10.3
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tabulate==0.9.0
tiktoken==0.4.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
unstructured==0.8.1
urllib3===1.24.2
uvicorn==0.24.0.post1
wcwidth==0.2.10
xlrd==2.0.1
XlsxWriter==3.1.9
yarl==1.9.2
